Initial commit
This is a first attempt at tracking an interactive todo list. It's currently built around the idea that you'll have items that repeat on a daily basis, so every day at midnight it will rotate out the previous day and in the current day. Completed todos are currently only stored as local storage on the browser, but should probably eventually be sent to another server somewhere that can coordinate the list of todos. This would more easily allow family members to add todos from their phone or similar.

Lots more to do here, including making the way it tracks daily items smarter. This is a good first step, though.